Check and propagate exit codes throughout job lifetime

+++ scripts/run_job.sh
@@ -65,14 +65,19 @@ echo "Running LHE and GEN+SIM step (`date`)";
 singularity run --home $PWD --bind /cvmfs $extra_bind --contain --ipc --pid $image \
   ./run_step0.sh $jobId $eventsPerLumi_nr $maxEvents_nr $era_nr $spin_nr $mass_nr  \
                  $decayMode_str $cmssw_host $cleanup_str step0;
+exit_code=$?;
 if [ ! -f step0.root ]; then
   echo "No output file was produced at step 0 -> exiting";
   exit 1;
 fi
+if [[ $exit_code -ne 0 ]]; then
+  exit $exit_code;
+fi;
 
 
 echo "Running PU premixing and AODSIM step (`date`)";
 ./run_step1.sh $jobId $era_nr $cmssw_host $cleanup_str step0 step1;
+exit_code=$?;
 if [ "$cleanup_str" == "true" ]; then
   rm -fv step0.root;
 fi;
@@ -80,9 +85,13 @@ if [ ! -f step1.root ]; then
   echo "No output file was produced at step 1 -> exiting";
   exit 1;
 fi
+if [[ $exit_code -ne 0 ]]; then
+  exit $exit_code;
+fi;
 
 echo "Running MiniAODSIM step (`date`)";
 ./run_step2.sh $era_nr $cmssw_host $cleanup_str step1 step2;
+exit_code=$?;
 if [ "$cleanup_str" == "true" ]; then
   rm -fv step1.root;
 fi;
@@ -90,13 +99,21 @@ if [ ! -f step2.root ]; then
   echo "No output file was produced at step 2 -> exiting";
   exit 1;
 fi
+if [[ $exit_code -ne 0 ]]; then
+  exit $exit_code;
+fi;
 
 if [ "$runNano_str" == "yes" ]; then
   ./run_step3.sh $era_nr $cmssw_host step2 step3;
+  exit_code=$?;
   mv -v step3.root tree.root;
 else:
   mv -v step2.root mini.root;
 fi;
+if [[ $exit_code -ne 0 ]]; then
+  exit $exit_code;
+fi;
+
 echo "All done (`date`)";
 
 ls -lh;
